refactor(back-office): move DataQuality components to `components` directory and update imports
